invariant testing

+++ test/Time.t.sol
@@ -6,16 +6,22 @@ import {Auction} from "../src/Time.sol";
 
 contract TimeTest is Test {
     Auction public auction;
-    uint public startTime;
+    uint256 public startTime;
 
     function setUp() public {
-        auction = new Auction();     
+        auction = new Auction();
         startTime = block.timestamp;
     }
 
     function testBid() public {
-       vm.expectRevert(bytes("cannot bid"));
-       vm.warp(block.timestamp+ 0.5 days );
-       auction.bid();
+        vm.expectRevert(bytes("cannot bid"));
+        vm.warp(block.timestamp + 0.5 days);
+        auction.bid();
     }
+
+    // function testTimestamp() public {
+    //     uint256 t = block.timestamp;
+    //     skip(100);
+    //     assertEq(block.timestamp == t + 100);
+    // }
 }
